HHH-14642 migrate idclass
