+++ hibernate-core/src/test/java/org/hibernate/orm/test/jpa/criteria/idclass/Tool.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.jpa.test.criteria.idclass;
+package org.hibernate.orm.test.jpa.criteria.idclass;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
